+++ tests/integration/test_issue_2299.py
@@ -1,7 +1,7 @@
 # Copyright 2025 Pex project contributors.
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from __future__ import print_function
+from __future__ import absolute_import
 
 import filecmp
 import glob